SF Bug #79 Template export from WinXP cannot import in Linux and Mac OS X
Use BINARY mode when reading and writing the template file.
Otherwise, the BraindeadOS puts some stray \r into the file.

I admit this is a beginners fault.

Add a fallback template load for windows to load
the template in ASCII mode if binary mode fails
to allow importing older templates on Windows.