Close #368: error while creating CRL with CA using EC key (ed25519)
This fix differs from the proposal in #368 because the signature
of pki_crl::sign() changed after the 2.4.0 release.

However, I was able to increase the bug impact!
Now it crashed instead of only throwing an error! :-)

The signature of the pki_crl::sign() method changed with
commit 61fa66b12976582d1f6f38df2308fd800be4cbef, but C++
"outsmarted" me and automatically transformed the EVP_MD == nullptr
into the digest::constructor(EVP_MD) which crashed on the nullptr.