Fix item search
Caching solution for 186657fbce4c9ff6a7c783b9f497d6e81424a6b0
was bad. It always returned "visible" when called with unchanged
pattern instead of the last result.  Cache the last result
in "iamvisible": 0=no, 1=yes, 2=one-of-my-children and return
the cached value as long as the search pattern is unchanged.