Translation: fix translation sources and xca.pot
After creating a library from the lib and widgets sources
to be used by xca and the test app, the source-collection
for xca does not work anymore. Collect them manually now.

xca.pot was erased during "make clean".
Create the xca.pot in the destination directory and copy it
to the source directory to avoid tagging "lang/xca.pot" as
"generated".