Close #532: XCA not in English on macos
The "en" language was missing from the list of languages
in Info.plist:CFBundleLocalization.
English is the source language for all translations and
not part of the LANGUAGES_PLIST variable.
The LANGUAGES_PLIST is compiled from the list of all
languages, XCA is translated to.

If no translated language is installed in the system,
the first one from the list will be picked, apparently.

Add static "en" entry to Info.plist:CFBundleLocalizations
as first language-entry and thus as fallback language if nothing
else matched.