CMake: fix dependency issue when creating documentation
Changing RST files in doc now result in a recreation
of the HTML and QtHelp files during build.

The sphinx-html and sphinx-qthelp targets need to
depend on sphinx-src to create that target files only once
(and not twice in parallel during parellel builds)
and also on all output files sphinx-src depends on.
.