Improve headerlist. Don't overload isNumeric
isNumeric now is a virtual function of the dbheader class
and overloaded by subclasses if needed.

The strict connection isNumeric == AlignLeft is removed
and a new virtual function alignment() is used to return the alignment.
By default this is "isNumeric() ? Qt::AlignRight : Qt::AlignLeft;"
but can now be overridden in sub-classes.

the key_dbheader sub-class vanished an got replaced by
the num_dbheader class with the capability to set the type to hd_key